[prepare-frontend-dev-env] - move frontend to docker, rename packages, refactor docker compose
