+++ backend/src/main/java/edu/issuetracker/Main.java
@@ -1,4 +1,4 @@
-package edu.st4rd.otus;
+package edu.issuetracker;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
