+++ backend/src/main/resources/application.yml
@@ -3,7 +3,7 @@ server:
 
 spring:
   datasource:
-    url: jdbc:postgresql://localhost:5433/issues_folder
+    url: jdbc:postgresql://localhost:5433/issue_tracker
     username: usr
     password: pwd
