+++ docker/.dockerignore
@@ -1,0 +1,7 @@
+.git
+.gitignore
+.vscode
+.angular
+docker-compose*.yml
+Dockerfile
+node_modules
