+++ docker/.env
@@ -1,0 +1,8 @@
+FRONTEND_PORT=4200
+KEYCLOAK_PORT=8484
+KEYCLOAK_POSTGRES_PORT=5432
+BACKEND_POSTGRES_PORT=5433
+PG_ADMIN_PORT=9090
+
+POSTGRES_USER=usr
+POSTGRES_PASSWORD=pwd
