+++ docker/backend-postgres/init_db.sh
@@ -1,0 +1,8 @@
+echo "=== Schema initialization for 'issue_tracker' database ==="
+psql issue_tracker usr -f /tmp/schema.sql
+
+echo "=== Data initialization for 'issue_tracker' database ==="
+for script in /tmp/data/*.sql;
+do
+    psql issue_tracker usr -f "$script"
+done
