+++ docker/docker-compose.yml
@@ -1,63 +1,108 @@
 version: "3.8"
 
 services:
-  postgres:
-    container_name: alpa-postgres
-    hostname: alpa-postgres
-    build: ./postgres
+
+  # docker compose up frontend
+  frontend:
+    container_name: frontend
+    hostname: frontend
+    build:
+      context: ../frontend
+      dockerfile: ../docker/frontend/Dockerfile
+    command: sh -c "npm start"
+    ports:
+      - "${FRONTEND_PORT}:4200"
+    volumes:
+      - node_modules:/app/node_modules
+      - ../frontend:/app
+    deploy:
+      resources:
+        limits:
+          cpus: '0.5'
+          memory: 1024M
+
+  # docker compose up backend-postgres
+  backend-postgres:
+    container_name: backend-postgres
+    hostname: backend-postgres
+    build: ./backend-postgres
     command: postgres -c max_connections=200 -p 5433
     environment:
-      POSTGRES_USER: ${POSTGRES_USER:-usr}
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-pwd}
-      POSTGRES_DB: issues_folder
+      POSTGRES_USER: ${POSTGRES_USER}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
+      POSTGRES_DB: issue_tracker
     ports:
-      - "5433:5433"
+      - "${BACKEND_POSTGRES_PORT}:5433"
     networks:
       - pgadmin
+    deploy:
+      resources:
+        limits:
+          cpus: '0.5'
+          memory: 512M
 
+  # docker compose up keycloak-postgres
   keycloak-postgres:
-    container_name: alpa-keycloak-postgres
-    hostname: alpa-keycloak-postgres
+    container_name: keycloak-postgres
+    hostname: keycloak-postgres
     image: library/postgres
     environment:
-      POSTGRES_USER: ${POSTGRES_USER:-usr}
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-pwd}
+      POSTGRES_USER: ${POSTGRES_USER}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
       POSTGRES_DB: keycloak_db
     ports:
-      - "5432:5432"
+      - "${KEYCLOAK_POSTGRES_PORT}:5432"
     restart: unless-stopped
     networks:
       - keycloak
       - pgadmin
+    deploy:
+      resources:
+        limits:
+          cpus: '0.5'
+          memory: 512M
 
+  # docker compose up keycloak
   keycloak:
-    container_name: alpa-keycloak
-    hostname: alpa-keycloak
+    container_name: keycloak
+    hostname: keycloak
     build: ./keycloak
     ports:
-      - "8484:8080"
+      - "${KEYCLOAK_PORT}:8080"
     depends_on:
       - keycloak-postgres
     networks:
       - keycloak
+    deploy:
+      resources:
+        limits:
+          cpus: '0.5'
+          memory: 1024M
 
+  # docker compose up pgadmin
   pgadmin:
-    container_name: alpa-pgadmin
-    hostname: alpa-pgadmin
+    container_name: pgadmin
+    hostname: pgadmin
     build: ./pgadmin
     environment:
-      - PGADMIN_DEFAULT_EMAIL=postgres@localhost
-      - PGADMIN_DEFAULT_PASSWORD=postgres
-      - PGADMIN_CONFIG_SERVER_MODE=False
-      - PGADMIN_CONFIG_MASTER_PASSWORD_REQUIRED=False
+      PGADMIN_DEFAULT_EMAIL: admin@dev.test
+      PGADMIN_DEFAULT_PASSWORD: admin
     ports:
-      - "9090:80"
+      - "${PG_ADMIN_PORT}:80"
     depends_on:
-      - postgres
+      - backend-postgres
       - keycloak-postgres
     networks:
       - pgadmin
+    deploy:
+      resources:
+        limits:
+          cpus: '0.5'
+          memory: 512M
 
 networks:
   keycloak:
   pgadmin:
+
+volumes:
+  node_modules:
