+++ docker/keycloak/realms/issue_tracker-realm.json
@@ -1,6 +1,6 @@
 {
-  "id": "issues_folder",
-  "realm": "issues_folder",
+  "id": "issue_tracker",
+  "realm": "issue_tracker",
   "notBefore": 0,
   "defaultSignatureAlgorithm": "RS256",
   "revokeRefreshToken": false,
@@ -47,7 +47,7 @@
     "realm": [
       {
         "id": "605f88b9-8b41-460b-a723-34f02c0323b2",
-        "name": "default-roles-issues_folder",
+        "name": "default-roles-issue_tracker",
         "description": "${role_default-roles}",
         "composite": true,
         "composites": {
@@ -63,7 +63,7 @@
           }
         },
         "clientRole": false,
-        "containerId": "issues_folder",
+        "containerId": "issue_tracker",
         "attributes": {}
       },
       {
@@ -72,7 +72,7 @@
         "description": "${role_offline-access}",
         "composite": false,
         "clientRole": false,
-        "containerId": "issues_folder",
+        "containerId": "issue_tracker",
         "attributes": {}
       },
       {
@@ -81,7 +81,7 @@
         "description": "${role_uma_authorization}",
         "composite": false,
         "clientRole": false,
-        "containerId": "issues_folder",
+        "containerId": "issue_tracker",
         "attributes": {}
       }
     ],
@@ -299,7 +299,7 @@
         }
       ],
       "security-admin-console": [],
-      "issues_folder_client": [],
+      "issue_tracker_client": [],
       "admin-cli": [],
       "account-console": [],
       "broker": [
@@ -397,11 +397,11 @@
   "groups": [],
   "defaultRole": {
     "id": "605f88b9-8b41-460b-a723-34f02c0323b2",
-    "name": "default-roles-issues_folder",
+    "name": "default-roles-issue_tracker",
     "description": "${role_default-roles}",
     "composite": true,
     "clientRole": false,
-    "containerId": "issues_folder"
+    "containerId": "issue_tracker"
   },
   "requiredCredentials": [
     "password"
@@ -463,7 +463,7 @@
       "disableableCredentialTypes": [],
       "requiredActions": [],
       "realmRoles": [
-        "default-roles-issues_folder"
+        "default-roles-issue_tracker"
       ],
       "notBefore": 0,
       "groups": []
@@ -493,13 +493,13 @@
       "clientId": "account",
       "name": "${client_account}",
       "rootUrl": "${authBaseUrl}",
-      "baseUrl": "/realms/issues_folder/account/",
+      "baseUrl": "/realms/issue_tracker/account/",
       "surrogateAuthRequired": false,
       "enabled": true,
       "alwaysDisplayInConsole": false,
       "clientAuthenticatorType": "client-secret",
       "redirectUris": [
-        "/realms/issues_folder/account/*"
+        "/realms/issue_tracker/account/*"
       ],
       "webOrigins": [],
       "notBefore": 0,
@@ -534,13 +534,13 @@
       "clientId": "account-console",
       "name": "${client_account-console}",
       "rootUrl": "${authBaseUrl}",
-      "baseUrl": "/realms/issues_folder/account/",
+      "baseUrl": "/realms/issue_tracker/account/",
       "surrogateAuthRequired": false,
       "enabled": true,
       "alwaysDisplayInConsole": false,
       "clientAuthenticatorType": "client-secret",
       "redirectUris": [
-        "/realms/issues_folder/account/*"
+        "/realms/issue_tracker/account/*"
       ],
       "webOrigins": [],
       "notBefore": 0,
@@ -658,7 +658,7 @@
     },
     {
       "id": "cdace2e3-7c5a-4985-b3f5-b9dfa33fde44",
-      "clientId": "issues_folder_client",
+      "clientId": "issue_tracker_client",
       "surrogateAuthRequired": false,
       "enabled": true,
       "alwaysDisplayInConsole": false,
@@ -710,7 +710,7 @@
         "profile",
         "roles",
         "email",
-        "issues_folder"
+        "issue_tracker"
       ],
       "optionalClientScopes": [
         "address",
@@ -761,13 +761,13 @@
       "clientId": "security-admin-console",
       "name": "${client_security-admin-console}",
       "rootUrl": "${authAdminUrl}",
-      "baseUrl": "/admin/issues_folder/console/",
+      "baseUrl": "/admin/issue_tracker/console/",
       "surrogateAuthRequired": false,
       "enabled": true,
       "alwaysDisplayInConsole": false,
       "clientAuthenticatorType": "client-secret",
       "redirectUris": [
-        "/admin/issues_folder/console/*"
+        "/admin/issue_tracker/console/*"
       ],
       "webOrigins": [
         "+"
@@ -1052,7 +1052,7 @@
     },
     {
       "id": "c06a85f8-217f-4d07-8105-21db10f0abd4",
-      "name": "issues_folder",
+      "name": "issue_tracker",
       "protocol": "openid-connect",
       "attributes": {
         "include.in.token.scope": "true",
