+++ docker/keycloak/scripts/commands.sh
@@ -1,3 +1,2 @@
-#!/bin/bash
 cd /opt/keycloak/bin
 ./kc.sh import --dir /tmp/realms --override true && ./kc.sh start-dev --db=postgres
